final commit

+++ TodoApp/Services/UserService.cs
@@ -1,11 +1,6 @@
 ﻿using SQLite;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using TodoApp.Data;
-using TodoApp.Models;
+using User = TodoApp.Models.User;
 
 namespace TodoApp.Services
 {
@@ -36,11 +31,20 @@ namespace TodoApp.Services
             return await _database.UpdateAsync(user);
         }
 
-        public async Task<int> DeleteUserAsync(User user)
+        public async Task<int> DeleteUserAsync(int userId)
         {
-            return await _database.DeleteAsync(user);
+            var user = await _database.Table<User>().Where(u => u.Id == userId).FirstOrDefaultAsync();
+            if (user != null)
+            {
+                return await _database.DeleteAsync(user);
+            }
+            return 0;
         }
 
+        public async Task<IEnumerable<User>> GetAllUsersAsync()
+        {
+            return await _database.Table<User>().ToListAsync();
+        }
         public async Task<User> AuthenticateUserAsync(string username, string password)
         {
             var user = await _database.Table<User>()
@@ -53,6 +57,12 @@ namespace TodoApp.Services
             }
 
             return user;
+        }
+
+        public async Task<User> GetUserByIdAsync(int userId)
+        {
+            return await _database.Table<User>()
+                .FirstOrDefaultAsync(u => u.Id == userId);
         }
 
         public void Logout()
